moved lib specs to lib folder
